Modified RestfulController to offer a protected AcceptHeader property, which allows for subclasses to obtain the request header from somewhere other than request.Headers["Accept"].
The reason for this: the latest Flash Player/Flex combinations do not reliably change the Accept request header when using URLRequestHeader or HTTPService. A workaround is to send the required Accept header under a pseudonym, e.g. "RestAccept". The server can test for the existance of this custom header first, and if not found, fall back to the actual Accept header.

git-svn-id: https://svn.castleproject.org/svn/castlecontrib/Castle.MonoRail.Rest/trunk@557 73e77b4c-caa6-f847-a29a-24ab75ae54b6